perusable/requirements.txt
bcrypt==3.2.0
black==22.6.0
email-validator==1.2.1
flake8==4.0.1
isort==5.10.1
pytest-cov==3.0.0
pytest-xdist==2.5.0
fastapi==0.78.0
passlib==1.7.4
pydantic==1.9.1
PyJWT==2.4.0
pytest==7.1.2
python-decouple==3.6
requests==2.28.1
starlette==0.19.1
uvicorn==0.18.2
